Issue 41864: DataRegion column analytics updates to lazy load Ext4 and vis dependencies - moves LABKEY.ColumnSummaryStatistics.showDialogFromDataRegion from ColumnSummaryStatistics.js to DataRegion.js - remove ColumnSummaryStatistics dependency and vis dependency from AnalyticsProvider classes - fix DataRegion update to column header menu item disabled state on add/remove column visualization - wrap ColumnVisualizationAnalytics.js functions in calls to LABKEY.requiresVisualization()
