+++ query/webapp/query/ColumnSummaryStatistics.js
@@ -4,44 +4,6 @@
  * Licensed under the Apache License, Version 2.0: http://www.apache.org/licenses/LICENSE-2.0
  */
 
-LABKEY.ColumnSummaryStatistics = new function () {
-    /**
-     * Used via SummaryStatisticsAnalyticsProvider to show a dialog of the applicable summary statistics for a column in the view.
-     * @param dataRegionName
-     * @param colFieldKey
-     */
-    var showDialogFromDataRegion = function(dataRegionName, colFieldKey) {
-        var region = LABKEY.DataRegions[dataRegionName];
-        if (region) {
-            var regionViewName = region.viewName || "",
-                column = region.getColumn(colFieldKey);
-
-            if (column) {
-                region.getColumnAnalyticsProviders(regionViewName, colFieldKey, function(colSummaryStats) {
-                    Ext4.create('LABKEY.ext4.ColumnSummaryStatisticsDialog', {
-                        queryConfig: region.getQueryConfig(),
-                        filterArray: LABKEY.Filter.getFiltersFromUrl(region.selectAllURL, 'query'), //Issue 26594
-                        containerPath: region.containerPath,
-                        column: column,
-                        initSelection: colSummaryStats,
-                        listeners: {
-                            applySelection: function(win, colSummaryStatsNames) {
-                                win.getEl().mask("Applying selection...");
-                                region.setColumnSummaryStatistics(regionViewName, colFieldKey, colSummaryStatsNames);
-                                win.close();
-                            }
-                        }
-                    }).show();
-                });
-            }
-        }
-    };
-
-    return {
-        showDialogFromDataRegion: showDialogFromDataRegion
-    };
-};
-
 Ext4.define('LABKEY.ext4.ColumnSummaryStatisticsDialog', {
     extend: 'Ext.window.Window',
