+++ visualization/resources/views/querySummary.view.xml
@@ -1,8 +1,7 @@
 <view xmlns="http://labkey.org/data/xml/view" title="Query Summary" frame="none">
     <dependencies>
         <dependency path="vis/vis"/>
-        <dependency path="vis/ColumnVisualizationAnalytics.js"/>
-        <dependency path="vis/ColumnVisualizationAnalytics.css"/>
+        <dependency path="vis/ColumnVisualizationAnalytics"/>
         <dependency path="query/ColumnSummaryStatistics"/>
     </dependencies>
 </view>
