+++ visualization/src/org/labkey/visualization/report/DimensionBarChartAnalyticsProvider.java
@@ -89,9 +89,7 @@ public class DimensionBarChartAnalyticsProvider extends ColumnAnalyticsProvider
     @Override
     public void addClientDependencies(Set<ClientDependency> dependencies)
     {
-        dependencies.add(ClientDependency.fromPath("vis/vis"));
-        dependencies.add(ClientDependency.fromPath("vis/ColumnVisualizationAnalytics.js"));
-        dependencies.add(ClientDependency.fromPath("vis/ColumnVisualizationAnalytics.css"));
+        dependencies.add(ClientDependency.fromPath("vis/ColumnVisualizationAnalytics"));
     }
 
     @Override
